pyproject.toml
[tool.poetry]
name = "login-test-o-parser"
version = "0.1.0"
description = ""
authors = ["Tim"]
readme = "README.md"
packages = [{include = "login_test_o_parser"}]

[tool.poetry.dependencies]
python = "^3.10"
fake-useragent = "^1.1.3"
beautifulsoup4 = "^4.12.2"
aiohttp = "^3.8.5"
selenium = "^4.10.0"
lxml = "^4.9.3"
django = "==3.*"
djangorestframework = "==3.*"
python-dotenv = "^1.0.0"
mysqlclient = "^2.2.0"
drf-yasg = "^1.21.7"
celery = "^5.3.1"
redis = "^4.6.0"
coreapi = "^2.3.3"
django-adminlte-3 = "^0.1.6"
aiogram = "==2.*"
aioredis = "^2.0.1"
isort = "^5.12.0"
webdriver-manager = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
